requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.5
urllib3==2.6.3
wheel==0.45.1
